-Tariff rate delete bug resolved

+++ Implementation/Services/TariffRateService.cs
@@ -233,6 +233,10 @@ namespace Cares.Implementation.Services
         /// <param name="standardRateMain"></param>
         public void DeleteTariffRate(StandardRateMain standardRateMain)
         {
+            if (standardRateMain.StandardRates.Count > 0)
+            {
+                throw new CaresException(string.Format(CultureInfo.InvariantCulture, Resources.Tariff.TariffRate.StRateIsAssociatedWithHireGroupError));
+            }
             standardRateMainRepository.Delete(standardRateMain);
             standardRateMainRepository.SaveChanges();
         }
